Promote moderately-recommended suggestions to highly-recommended
This will happen when a moderately-recommended suggestion (one that
has a matching suggestion in the other direction for every one of its
letters) has matches that are, themselves, moderately-recommended. In
other words, a highly-recommended suggestion can be used since it will
definitely have a usable set of suggestions in the opposite direction.