Optimize sort order.
Sort primarily by increasing number of candidate answers and then by
decreasing length of answer.

The theory is that that will give us a strategy to search through the
fewest number of candidates first and, within that, visit the answers
that have the most impact on the puzzle (the longest answers).